Allow option to remove host tag for dogstatsd metrics
This change adds a config option (dogstatsd_remove_host_tag) which when
enabled will force the dogstatsd daemon to report all metrics with an
added host: tag. This is done in an effort to reduce the cardinality on
custom metrics at the expense of being able to break the metrics down at
the host level.

Note that this was built from the 5.19.0 tag.