Implement basic blacklist for tags sent via dogstatsd
This change adds a config option (`dogstatsd_blacklist_tags_re`) to be
used as a regular expression that can filter out unwanted metric tags
received via `dogstatsd`. This is another change meant to allow more
flexibility in reducing the cardinality on custom metrics.

Note that this was built from the 5.19.0 tag.